requirements.txt
# MCP dependencies
mcp>=1.3.0
fastmcp>=0.4.1

# Web server
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0

# HTTP requests and data processing
requests>=2.28.2
pandas>=2.0.0
xmltodict>=0.13.0
beautifulsoup4>=4.12.0

# Environment and configuration
python-dotenv>=1.0.0
argparse>=1.4.0

# For deployment
gunicorn>=20.1.0

# Testing (optional)
# pytest>=7.3.1
# pytest-cov>=4.1.0
